llm/requirements.txt
boto3==1.35.68
fastapi==0.115.5
python-dotenv==1.0.1
uvicorn==0.32.1
instructor==1.6.4
anthropic[bedrock]==0.39.0
python-multipart==0.0.17
pymupdf==1.24.14
openpyxl==3.1.5
pandas==2.2.3
python-docx==1.1.2
libsql-experimental==0.0.41